moved searchresults outside of navbar
