+++ my-app/src/Searchbar.js
@@ -3,38 +3,30 @@ import {useRef } from "react";
 import FoodList from "./HomePage/FoodList";
 var hidden='none'
 const Searchbar = (props) => {
-    const [searchInput, setSearchInput] = useState("");
+    const setSearchInput = props.setSearchInput;
     var items = props.items;
     items = items.map((item)=>{return {name : item.name.toUpperCase(), category: item.category};});
    //  console.log(items);
     const handleChange = (e) => {
        e.preventDefault();
        setSearchInput(e.target.value)
-       hidden = "block"
-       console.log(hidden)
-       console.log(e.target.value)
     };
-    function searchList() {
-       let filtereditems = items.filter((item) => {
-          return item.name.match(searchInput.toUpperCase());
-       });
-       const filtered = filtereditems?.map((item) => (
-          <tr>
-             <a href = {'/Categories/CategoryFoodList-' + item.category}> {item.name}</a>
-          </tr>
-       ));
-       if(searchInput==="") return ;
-       return ( 
-            <div>
-                <tr>
-                {/* <th> Item </th>
-                <th></th>
-                <th> Cost </th> */}
-                </tr>
-                <div> {filtered} </div>
-            </div>
-        );
-    }
+   //  function searchList() {
+   //     let filtereditems = items.filter((item) => {
+   //        return item.name.match(searchInput.toUpperCase());
+   //     });
+   //     const filtered = filtereditems?.map((item) => (
+   //        <tr>
+   //           <a href = {'/Categories/CategoryFoodList-' + item.category}> {item.name}</a>
+   //        </tr>
+   //     ));
+   //     if(searchInput==="") return ;
+   //     return ( 
+   //          <div>
+   //              <div> {filtered} </div>
+   //          </div>
+   //      );
+   //  }
 
     const myref = useRef(null);
     function scrollToCategory(CategoryId) {
@@ -46,13 +38,12 @@ const Searchbar = (props) => {
              type="search"
              placeholder="Search..."
              onChange={handleChange}
-             value={searchInput}
              className="search"
           />
-          <table style={{ tableLayout: "fixed", width: "11rem", border:"none"}} className="searchresults" >
+          {/* <table style={{ tableLayout: "fixed", width: "11rem", border:"none"}} className="searchresults" >
              
              {searchList()}
-          </table>
+          </table> */}
        </div>
     );
  }
